requirements.txt
# Satori Lite Requirements
# Minimal dependencies for lightweight Satori neuron
# Python 3.12+ required

# Core dependencies
setuptools>=65.0.0

# Data handling - pinned for Python 3.12 compatibility
numpy==1.26.4
pandas==2.2.3
pyarrow==18.1.0
fastparquet==2024.11.0

# Configuration
PyYAML==6.0.2

# System utilities
psutil==6.1.0
joblib==1.4.2

# AI Engine - Machine Learning (pinned for Python 3.12)
xgboost==2.1.3
scikit-learn==1.5.2
lightgbm==4.5.0
catboost==1.2.7
statsmodels==0.14.4
skforecast==0.14.0
pmdarima==2.0.4
linear-tree==0.3.5
sktime==0.35.0
optuna==4.1.0
# ppscore==1.3.0  # Removed: abandoned since 2021, requires pandas<2.0

# AI Engine - Deep Learning (optional, for advanced models)
# torch>=2.3.1  # typically installed in docker image
transformers==4.47.0
# prophet>=1.1.6  # has complex dependencies
# tbats>=1.1.3
# neuralforecast>=1.7.5

# Lomb-scargle interpolation
astropy==7.0.0

# Wallet dependencies - Ravencoin and Evrmore blockchain
python-ravencoinlib
python-evrmorelib
mnemonic==0.21
base58==2.1.1

# Encryption for wallet vault
pycryptodome==3.21.0

# Ethereum wallet support
eth-account==0.13.4
eth-keys==0.6.1
web3>=7.0.0
coincurve>=20.0.0

# API validation
marshmallow==3.23.1

# Networking
requests==2.32.3
websockets>=10.0.0
trio==0.27.0
trio-websocket==0.11.1

# Web UI
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.4.1
# simple-websocket provides WebSocket support for Flask-SocketIO threading mode
simple-websocket>=1.0.0
# gunicorn for production-ready WSGI serving (required, no fallback)
gunicorn>=21.0.0
qrcode[pil]==8.0

# Data encryption
cryptography==44.0.0

# IP address validation
ipaddress>=1.0.0

# Centrifugo real-time messaging
centrifuge-python==0.4.1

# Optional: Telemetry logging
# Uncomment lines below to enable telemetry
# sanic>=23.6.0
# aiosqlite>=0.19.0
# uvloop>=0.19.0

# Optional: P2P Networking
# Enables hybrid and pure P2P modes (set SATORI_NETWORKING_MODE=hybrid or p2p)
# See: https://github.com/SatoriNetwork/satorip2p
satorip2p>=1.0.0
